Abstracting utils
Added utils to utils.js
Added jsdocs
Imported utils in game.js
Removed utils from game.js
